Tech/refactor email prepare (#2852)
* Use ToRecepientData instead of ctor in two more places

* TelegramId should be long

* add new ctor for RecepientData

+++ src/JoinRpg.PrimitiveTypes/TelegramId.cs
@@ -1,5 +1,5 @@
 namespace JoinRpg.PrimitiveTypes;
 
-public record TelegramId(int Id, PrefferedName? UserName)
+public record TelegramId(long Id, PrefferedName? UserName)
 {
 }
